add userAuth.controller
